+++ packages/data-types/src/index.ts
@@ -94,3 +94,4 @@ export * from './lib/types/calendar-event.type';
 export * from './lib/types/minimal-npc.type';
 export * from './lib/types/gifting-preference-keys.type';
 export * from './lib/types/minimal-tag-based-item.type';
+export * from './lib/types/available-languages.type';