+++ packages/data-types/src/lib/interfaces/offering-altar.interface.ts
@@ -2,6 +2,7 @@ import { Offerings } from "./offerings.interface";
 
 export interface OfferingAltar {
     key: string;
+    urlPath: string;
     offeringGroupTitle: string;
     offeringGroupRewardText: string;
     offerings: Offerings[]