+++ packages/guide/src/app/my-coral-guide/components/cooking-recipes-checklist/cooking-recipes-checklist.component.ts
@@ -10,7 +10,18 @@ export class CookingRecipesChecklistComponent extends BaseItemChecklistComponent
     checklistService = inject(CookingRecipesChecklistService);
     checklistDefinition$ = this._database.fetchCookingRecipesChecklist$();
 
+    protected cookingUtensilMapping = this._database.getCookingUtensilMapping()
     constructor() {
         super();
+    }
+
+    override urlPathFromLabel = (label: string) => {
+
+        const foundKey = Object.keys(this.cookingUtensilMapping).find(key => this.cookingUtensilMapping[key].displayName === label);
+        if (foundKey) {
+            return foundKey
+        }
+
+        return label.toLowerCase().replaceAll(' ', '')
     }
 }