+++ packages/guide/src/app/shared/interfaces/settings.interface.ts
@@ -1,3 +1,7 @@
+import { AvailableLanguage } from "@ci/data-types";
+
 export interface Settings {
+    version: number;
     useBeta: boolean;
+    language: AvailableLanguage;
 }