+++ packages/pak-files-parser/src/util/string-table.class.ts
@@ -1,13 +1,17 @@
 import { readAsset } from "./functions";
 import { RawStringTable } from "../interfaces/raw-data-interfaces/raw-string-table.interface";
 import { SourceString } from "../types/source-string.type";
+import { AvailableLanguage } from "@ci/data-types";
 
 export class StringTable {
 
+    static defaultLang: AvailableLanguage = 'en';
     private static readonly cachedStringTables: Map<string, RawStringTable> = new Map<string, RawStringTable>()
     private static readonly cachedLocalizationFiles: Map<string, Record<string, Record<string, string>>> = new Map<string, Record<string, Record<string, string>>>()
 
-    public static getString(ref: SourceString, lang: 'de' | 'en' | 'es' | 'fr' | 'id' | 'ja' | 'zh-CN' = 'en'): string | null {
+    public static getString(ref: SourceString, lang?: AvailableLanguage): string | null {
+        if (!lang) lang = StringTable.defaultLang;
+
         if ('TableId' in ref) {
             const stringTablePath = ref.TableId.replace('/Game/ProjectCoral/', '/ProjectCoral/Content/ProjectCoral/').split('.')[0] + '.json';