+++ packages/guide/CHANGELOG.md
@@ -4,12 +4,15 @@
 
 #### Added
 
-- Added B.O.S., Bens Caravan and Socket and Pan shop items
+- WIP: Added translation of in-game data. Text added by Coral Guide won't be translated in this version.
+    - Go to "my guide" -> "settings" to change your language.
+- Added B.O.S., Bens caravan and Socket and Pan shop items
 - Added manual mapping for cooking utensils to show in database to reduce confusion about internal names
 - Added more images for generic items like "any butter" etc.
 
 #### Changed
 
+- Fixed checklist tab names being from in some cases
 - Fixed altars and offering checklist showing wrong item instead of tag based item
 - Fixed Joko and Valentina head image being broken
 - Fixed a few descriptions being missing